kitchensync_flask/requirements.txt

flask==2.1.0
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
flask-bcrypt==0.7.1
python-dotenv==0.19.2
PyJWT==2.3.0
